requirements.txt
streamlit==1.32.2
pandas==2.2.2
scikit-learn==1.4.2
matplotlib==3.8.4
seaborn==0.13.2
sentence-transformers==2.2.2
faiss-cpu==1.7.4
PyMuPDF==1.23.22
